+++ packages/lib/src/model/prop.ts
@@ -7,13 +7,19 @@ import { IsOptionalValue } from "../utils/types"
 export const noTypeChecker = Symbol("noTypeChecker")
 
 /**
+ * @ignore
+ */
+export const noDefaultValue = Symbol("noDefaultValue")
+
+/**
  * A model property.
  */
-export interface ModelProp<TValue, THasDefault> {
+export interface ModelProp<TValue, TCreationValue, TIsOptional> {
   $valueType: TValue
-  $hasDefault: THasDefault
-  defaultFn?: () => TValue
-  defaultValue?: TValue
+  $creationValueType: TCreationValue
+  $isOptional: TIsOptional
+  defaultFn: (() => TValue) | typeof noDefaultValue
+  defaultValue: TValue | typeof noDefaultValue
   typeChecker: TypeChecker | LateTypeChecker | typeof noTypeChecker
 }
 
@@ -21,15 +27,19 @@ export interface ModelProp<TValue, THasDefault> {
  * Model properties.
  */
 export interface ModelProps {
-  [k: string]: ModelProp<any, any>
+  [k: string]: ModelProp<any, any, any>
 }
 
 export type OptionalModelProps<MP extends ModelProps> = {
-  [K in keyof MP]: (MP[K]["$hasDefault"] & K)
+  [K in keyof MP]: (MP[K]["$isOptional"] & K)
 }[keyof MP]
 
 export type ModelPropsToData<MP extends ModelProps> = {
   [k in keyof MP]: MP[k]["$valueType"]
+}
+
+export type ModelPropsToCreationData<MP extends ModelProps> = {
+  [k in keyof MP]: MP[k]["$creationValueType"]
 }
 
 /**
@@ -44,11 +54,11 @@ export type ModelPropsToData<MP extends ModelProps> = {
  * @typeparam TValue Value type.
  * @returns
  */
-export function prop<TValue>(): ModelProp<TValue, IsOptionalValue<TValue, string, never>>
+export function prop<TValue>(): ModelProp<TValue, TValue, IsOptionalValue<TValue, string, never>>
 
 /**
  * Defines a model property, with an optional function to generate a default value
- * if not present.
+ * if the input snapshot / model creation data is `null` or `undefined`.
  *
  * Example:
  * ```ts
@@ -60,11 +70,13 @@ export function prop<TValue>(): ModelProp<TValue, IsOptionalValue<TValue, string
  * @param defaultFn Default value generator function.
  * @returns
  */
-export function prop<TValue>(defaultFn: () => TValue): ModelProp<TValue, string>
+export function prop<TValue>(
+  defaultFn: () => TValue
+): ModelProp<TValue, TValue | null | undefined, string>
 
 /**
  * Defines a model property, with an optional default value
- * if not present.
+ * if the input snapshot / model creation data is `null` or `undefined`.
  * You should only use this with primitive values and never with object values
  * (array, model, object, etc).
  *
@@ -77,15 +89,20 @@ export function prop<TValue>(defaultFn: () => TValue): ModelProp<TValue, string>
  * @param defaultValue Default primitive value.
  * @returns
  */
-export function prop<TValue>(defaultValue: TValue): ModelProp<TValue, string>
+export function prop<TValue>(
+  defaultValue: TValue
+): ModelProp<TValue, TValue | null | undefined, string>
 
-export function prop<TValue>(def?: any): ModelProp<TValue, any> {
+export function prop<TValue>(def?: any): ModelProp<TValue, any, any> {
+  const hasDefaultValue = arguments.length > 0
   const isDefFn = typeof def === "function"
+
   return {
     $valueType: null as any,
-    $hasDefault: null as any,
-    defaultFn: isDefFn ? def : undefined,
-    defaultValue: isDefFn ? undefined : def,
+    $creationValueType: null as any,
+    $isOptional: null as any,
+    defaultFn: hasDefaultValue && isDefFn ? def : noDefaultValue,
+    defaultValue: hasDefaultValue && !isDefFn ? def : noDefaultValue,
     typeChecker: noTypeChecker,
   }
 }