+++ packages/lib/test/snapshot/snapshot.test.ts
@@ -448,7 +448,7 @@ test("types", () => {
   assert(
     _ as SnapshotInOf<P2>,
     _ as ({
-      y?: number
+      y?: number | null
     } & {
       [modelTypeKey]: string
       [modelIdKey]: string
@@ -468,8 +468,8 @@ test("types", () => {
   assert(
     _ as SnapshotInOf<P>,
     _ as ({
-      x?: number
-      arr?: number[]
+      x?: number | null
+      arr?: number[] | null
       p2?: SnapshotInOf<P2>
     } & {
       [modelTypeKey]: string